use OBS and TARGET in spectrum extraction
