+++ gammapy/background/background_estimate.py
@@ -13,6 +13,7 @@ class BackgroundEstimate(object):
     """Cointainer class for background estimate
 
     This is meant to hold the result from a region based background estimation
+    for one observation. 
 
     Parameters:
     -----------
@@ -32,17 +33,15 @@ class BackgroundEstimate(object):
         self.alpha = alpha
 
 
-def ring_background_estimate(target, events, inner_radius, outer_radius):
+def ring_background_estimate(pos, on_radius, inner_radius, outer_radius, events):
     """Simple ring background estimate
 
     No acceptance correction is applied
     """
-    pos = target.on_region.center
-    radius = target.on_region.radius
     off_events = events.select_sky_ring(pos, inner_radius, outer_radius)
     # TODO : Replace with AnnulusSkyRegion
     off_region = dict(inner=inner_radius, outer=outer_radius)
-    alpha = ring_area_factor(radius, inner_radius, outer_radius)
+    alpha = ring_area_factor(on_radius, inner_radius, outer_radius)
 
     return BackgroundEstimate(off_region, off_events, alpha, tag='ring')
