+++ gammapy/spectrum/core.py
@@ -1,15 +1,18 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from ..utils.nddata import NDDataArray, BinnedDataAxis
 from astropy.table import Table
+from ..utils.nddata import NDDataArray, BinnedDataAxis
+from ..data import EventList
+from ..extern.pathlib import Path
+import astropy.units as u
 import numpy as np
 
 __all__ = [
     'CountsSpectrum',
-    #'OnCountsSpectrum',
-    #'OffCountsSpectrum',
+    'OnCountsSpectrum',
     'SpectrumObservation',
+    'SpectrumObservationList',
 ]
 
 class CountsSpectrum(NDDataArray):
@@ -63,7 +66,7 @@ class CountsSpectrum(NDDataArray):
         bin_lo = self.energy.data[:-1]
         bin_hi = self.energy.data[1:]
         names = ['CHANNEL', 'COUNTS', 'BIN_LO', 'BIN_HI']
-        meta = dict(name='SPECTRUM', hduclass='OGIP', hduclas1 = 'SPECTRUM')
+        meta = dict()
         return Table([channel, counts, bin_lo, bin_hi], names=names, meta=meta)
 
     def fill(self, events):
@@ -75,12 +78,12 @@ class CountsSpectrum(NDDataArray):
             List of event energies
         """
 
-        if isinstance(events, gammapy.data.EventList):
+        if isinstance(events, EventList):
             events = events.energy
 
         energy = events.to(self.energy.unit)
         binned_val = np.histogram(energy.value, self.energy.data.value)[0]
-        self.data = binned_val 
+        self.data = binned_val * u.ct 
 
     @property
     def total_counts(self):
@@ -168,4 +171,375 @@ class CountsSpectrum(NDDataArray):
         return cls(counts.decompose(), e_reco)
 
 
-class SpectrumObservation()
+class OnCountsSpectrum(CountsSpectrum):
+    """OGIP PHA equivalent
+    
+    Parameters
+    ----------
+    data : `~numpy.array`, list
+        Counts
+    energy : `~gammapy.utils.energy.EnergyBounds`
+        Bin edges of energy axis
+    livetime : `~astropy.units.Quantity`
+        Observation live time
+    obs_id : int
+        Unique identifier
+    """
+    def __init__(self, **kwargs):
+        super(CountsSpectrum, self).__init__(**kwargs)
+        self.phafile = 'pha_obs{}.fits'.format(self.obs_id)
+        self.arffile = self.phafile.replace('pha', 'arf')
+        self.rmffile = self.phafile.replace('pha', 'rmf')
+        self.bkgfile = self.phafile.replace('pha', 'bkg')
+
+    def to_table(self):
+        """Write"""
+        table = super(OnCountsSpectrum, self).to_table()
+        meta = dict(name='SPECTRUM', hduclass='OGIP', hduclas1 = 'SPECTRUM',
+                    backfile=self.bkgfile, respfile=self.rmffile,
+                    ancrfile=self.arffile )
+        table.meta = meta
+        return table
+
+
+class SpectrumObservation(object):
+    """1D spectral analysis storage class
+    
+    This container holds the ingredients for 1D region based spectral analysis
+    TODO: describe PHA, ARF, etc.
+
+    Meta data is stored in the ``on_vector`` attribute. This reflects the OGIP
+    convention.
+
+    Parameters
+    ----------
+    on_vector : `~gammapy.spectrum.OnCountsSpectrum`
+        On vector
+    off_vector : `~gammapy.spectrum.CountsSpectrum`
+        Off vector
+    arf : `~gammapy.irg.EffectiveAreaTable`
+        Effective Area
+    edisp : `~gammapy.irf.EnergyDispersion`
+        Energy dispersion matrix
+    """
+
+    def __init__(self, on_vector, off_vector, arf, edisp):
+        self.on_vector = on_vector
+        self.off_vector = off_vector
+        self.arf = arf
+        self.edisp = edisp
+
+    @property
+    def obs_id(self):
+        return self.on_vector.obs_id
+
+    @classmethod
+    def read(cls, phafile):
+        """Read `~gammapy.spectrum.SpectrumObservation` from OGIP files.
+
+        BKG file, ARF, and RMF must be set in the PHA header.
+
+        Parameters
+        ----------
+        phafile : str
+            OGIP PHA file to read
+        """
+        # Put here due to circular imports issues
+        from ..irf import EnergyDispersion, EffectiveAreaTable
+
+        f = make_path(phafile)
+        base = f.parent
+        on_vector = CountsSpectrum.read_pha(f)
+        meta = on_vector.meta
+        energy_dispersion = EnergyDispersion.read(str(base / meta.rmf))
+        effective_area = EffectiveAreaTable.read(str(base / meta.arf))
+        off_vector = CountsSpectrum.read_bkg(str(base / meta.bkg),
+                                             str(base / meta.rmf))
+
+        meta.update(phafile=phafile)
+        return cls(on_vector, off_vector, energy_dispersion, effective_area,
+                   meta)
+
+    def write(self, outdir=None, overwrite=True):
+        """Write OGIP files
+
+        The files are given default names and referenced in
+        the :ref:`gadf:ogip-pha` FITS header. 
+
+        Parameters
+        ----------
+        outdir : `~gammapy.extern.pathlib.Path`
+            output directory, default: pwd
+        overwrite : bool
+            Overwrite
+        """
+
+        outdir = Path(outdir) or Path.cwd()
+        outdir.mkdir(exist_ok=True, parents=True)
+
+        phafile = self.on_vector.phafile
+        bkgfile = self.on_vector.bkgfile
+        arffile = self.on_vector.arffile
+        rmffile = self.on_vector.rmffile
+
+        self.on_vector.write(outdir / phafile, overwrite=overwrite)
+        self.off_vector.write(outdir / bkgfile, overwrite=overwrite)
+        # Write in keV and cm2
+        self.arf.data = self.arf.data.to('cm2')
+        self.arf.energy.data = self.arf.energy.data.to('keV')
+        self.arf.write(outdir/arffile, overwrite=overwrite)
+        self.edisp.write(str(outdir / rmffile), energy_unit='keV',
+                                     clobber=overwrite)
+
+    @property
+    def excess_vector(self):
+        """Excess vector
+        Excess = n_on - alpha * n_off
+        """
+        return self.on_vector + self.off_vector * self.alpha * -1
+
+    @property
+    def spectrum_stats(self):
+        """`~gammapy.spectrum.results.SpectrumStats`
+        """
+        n_on = self.on_vector.total_counts
+        n_off = self.off_vector.total_counts
+        val = dict()
+        val['n_on'] = n_on
+        val['n_off'] = n_off
+        val['alpha'] = self.alpha
+        val['excess'] = float(n_on) - float(n_off) * self.alpha
+        val['energy_range'] = self.meta.energy_range
+        return SpectrumStats(**val)
+
+    def apply_energy_cut(self, energy_range=None, method='binned'):
+        """Restrict to a given energy range
+
+        If no energy range is given, it will be extracted from the PHA header.
+        Tow methods are available . Unbinned method: The new counts vectors are
+        created from the list of on and off events. Therefore this list must be
+        saved in the meta info. Binned method: The counts vectors are taken as
+        a basis for the energy range restriction. Only bins that are entirely
+        contained in the desired energy range are copied.
+
+        Parameters
+        ----------
+        energy_range : `~gammapy.utils.energy.EnergyBounds`, optional
+            Desired energy range
+        method : str {'unbinned', 'binned'}
+            Use unbinned on list / binned on vector
+
+        Returns
+        -------
+        obs : `~gammapy.spectrum.spectrum_extraction.SpectrumObservation`
+            Spectrum observation in desired energy range
+        """
+
+        if energy_range is None:
+            arf = self.effective_area
+            energy_range = [arf.energy_thresh_lo, arf.energy_thresh_hi]
+
+        energy_range = EnergyBounds(energy_range)
+        ebounds = self.on_vector.energy_bounds
+        if method == 'unbinned':
+            on_list_temp = self.meta.on_list.select_energy(energy_range)
+            off_list_temp = self.meta.off_list.select_energy(energy_range)
+            on_vec = CountsSpectrum.from_eventlist(on_list_temp, ebounds)
+            off_vec = CountsSpectrum.from_eventlist(off_list_temp, ebounds)
+        elif method == 'binned':
+            val = self.on_vector.energy_bounds.lower_bounds
+            mask = np.invert(energy_range.contains(val))
+            on_counts = np.copy(self.on_vector.counts)
+            on_counts[mask] = 0
+            off_counts = np.copy(self.off_vector.counts)
+            off_counts[mask] = 0
+            on_vec = CountsSpectrum(on_counts, ebounds)
+            off_vec = CountsSpectrum(off_counts, ebounds)
+        else:
+            raise ValueError('Undefined method: {}'.format(method))
+
+        off_vec.meta.update(backscal=self.off_vector.meta.backscal)
+        m = copy.deepcopy(self.meta)
+        m.update(energy_range=energy_range)
+
+        return SpectrumObservation(self.obs_id, on_vec, off_vec,
+                                   self.energy_dispersion, self.effective_area,
+                                   meta=m)
+
+    def _check_binning(self, **kwargs):
+        """Check that ARF and RMF binnings are compatible
+        """
+        raise NotImplementedError
+
+
+class SpectrumObservationList(list):
+    """
+    List of `~gammapy.spectrum.SpectrumObservation`.
+    """
+
+    def get_obslist_from_ids(self, id_list):
+        """Return an subset of the observation list
+
+        Parameters
+        ----------
+        id_list : list of int
+            List of Observation Id (runnumber)
+
+        Returns
+        -------
+        observation : `~gammapy.spectrum.SpectrumObservationList`
+            Subset of observations
+        """
+        new_list = list()
+
+        for id in id_list:
+            ids = [o.meta.obs_id for o in self]
+            try:
+                i = ids.index(id)
+            except ValueError:
+                raise ValueError("Observation {} not in list".format(id))
+
+            new_list.append(self[i])
+
+        return SpectrumObservationList(new_list)
+
+    @property
+    def total_spectrum(self):
+        """Stack all observations belongig to the list"""
+        return SpectrumObservation.stack_observation_list(self)
+
+    def info(self):
+        """Info string"""
+        ss = " *** SpectrumObservationList ***"
+        ss += "\n\nNumber of observations: {}".format(len(self))
+        ss += "\nObservation IDs: {}".format([o.obs_id for o in self])
+
+        return ss
+
+    def filter_by_reflected_regions(self, n_min):
+        """Filter observation list according to number of reflected regions.
+
+        Condition: number of reflected regions >= nmin.
+
+        Parameters
+        ----------
+        n_min : int
+            Minimum number of reflected regions
+
+        Returns
+        -------
+        idx : `~np.array`
+            Indices of element fulfilling the condition
+        """
+        val = [o.off_vector.meta.backscal for o in self]
+        condition = np.array(val) >= n_min
+        idx = np.nonzero(condition)
+        return idx[0]
+
+    def write(self, outdir=None, **kwargs):
+        """Create OGIP files
+
+        Parameters
+        ----------
+        outdir : str, `~gammapy.extern.pathlib.Path`, optional
+            Output directory, default: pwd
+        """
+        for obs in self:
+            obs.write(outdir=outdir, **kwargs)
+
+    @classmethod
+    def from_observation_table(cls, obs_table):
+        """Create `~gammapy.spectrum.SpectrumObservationList` from an
+        observation table.
+
+        Parameters
+        ----------
+        obs_table : `~gammapy.data.ObservationTable`
+            Observation table with column ``PHAFILE``
+        """
+        obs = [SpectrumObservation.read_ogip(_) for _ in obs_table['PHAFILE']]
+
+        return cls(obs)
+
+
+def stack(cls, specobs, group_id=None):
+        """Stack `~gammapy.spectrum.SpectrumObservation`
+
+        Observation stacking is implemented as follows
+        Averaged exposure ratio between ON and OFF regions, arf and rmf
+        :math:`\\alpha_{\\mathrm{tot}}`  for all observations is calculated as
+        .. math:: \\alpha_{\\mathrm{tot}} = \\frac{\\sum_{i}\\alpha_i \\cdot N_i}{\\sum_{i} N_i}
+        .. math:: \\arf_{\\mathrm{tot}} = \\frac{\\sum_{i}\\arf_i \\cdot \\livetime_i}{\\sum_{i} \\livetime_i}
+        .. math:: \\rmf_{\\mathrm{tot}} = \\frac{\\sum_{i}\\rmf_i \\cdot arf_i \\cdot livetime_i}{\\sum_{i} arf_i \\cdot livetime_i}
+
+        Parameters
+        ----------
+        obs_list : list of `~gammapy.spectrum.SpectrumObservations`
+            Observations to stack
+        group_id : int, optional
+            ID for stacked observations
+    
+        Returns
+        ------- 
+        stacked_obs : `~gammapy.spectrum.SpectrumObservations`
+        """
+
+        group_id = obs_list[0].meta.obs_id if group_id is None else group_id
+
+        # Stack ON and OFF vector using the _add__ method in the CountSpectrum class
+        on_vec = np.sum([o.on_vector for o in obs_list])
+
+        # If obs_list contains only on element np.sum does not call the
+        #  _add__ method which lead to a faulty meta object
+        if len(obs_list) == 1:
+            on_vec.meta = Bunch(livetime=obs_list[0].meta.livetime,
+                                backscal=1)
+
+        on_vec.meta.update(obs_id=group_id)
+
+        off_vec = np.sum([o.off_vector for o in obs_list])
+
+        # Stack arf vector
+        arf_band = [o.effective_area.data * o.meta.livetime.value for o in obs_list]
+        arf_band_tot = np.sum(arf_band, axis=0)
+        livetime_tot = np.sum([o.meta.livetime.value for o in obs_list])
+        arf_vec = arf_band_tot / livetime_tot
+        energy = obs_list[1].effective_area.energy.data
+        data = arf_vec * obs_list[0].effective_area.data.unit
+
+        arf = EffectiveAreaTable(energy=energy, data=data) 
+                                 
+        # Stack rmf vector
+        rmf_band = [o.energy_dispersion.pdf_matrix.T * o.effective_area.data.value * o.meta.livetime.value for
+                    o in obs_list]
+        rmf_band_tot = np.sum(rmf_band, axis=0)
+        pdf_mat = rmf_band_tot / arf_band_tot
+        etrue = obs_list[0].energy_dispersion.true_energy
+        ereco = obs_list[0].energy_dispersion.reco_energy
+        inan = np.isnan(pdf_mat)
+        pdf_mat[inan] = 0
+        rmf = EnergyDispersion(pdf_mat.T, etrue, ereco)
+
+        # Calculate average alpha
+        alpha_band = [o.alpha * o.off_vector.total_counts for o in obs_list]
+        alpha_band_tot = np.sum(alpha_band)
+        off_tot = np.sum([o.off_vector.total_counts for o in obs_list])
+        alpha_mean = alpha_band_tot / off_tot
+        off_vec.meta.backscal = 1. / alpha_mean
+
+        # Calculate energy range
+        # TODO: for the moment we take the minimal safe energy range
+        # Taking the whole range requires an energy dependent lifetime
+        emin = max([_.meta.safe_energy_range[0] for _ in obs_list])
+        emax = min([_.meta.safe_energy_range[1] for _ in obs_list])
+
+        m = Bunch()
+        m['energy_range'] = EnergyBounds([emin, emax])
+        on_vec.meta['safe_energy_range'] = EnergyBounds([emin, emax])
+        # m['safe_energy_range'] = EnergyBounds([emin, emax])
+        m['obs_ids'] = [o.meta.obs_id for o in obs_list]
+        m['alpha_method1'] = alpha_mean
+        m['livetime'] = Quantity(livetime_tot, "s")
+        m['group_id'] = group_id
+
+        return cls(on_vec, off_vec, rmf, arf, meta=m)
